Refactor packages from `endpoint` and `upstream` to `server`
