+++ core/src/main/java/dev/gihwan/tollgate/core/server/UpstreamFactory.java
@@ -22,23 +22,35 @@
  * SOFTWARE.
  */
 
-package dev.gihwan.tollgate.core.endpoint;
+package dev.gihwan.tollgate.core.server;
 
-import dev.gihwan.tollgate.core.upstream.UpstreamFactory;
-import dev.gihwan.tollgate.core.upstream.UpstreamService;
+import static java.util.Objects.requireNonNull;
 
-final class DefaultEndpointService implements EndpointService {
+import java.util.HashMap;
+import java.util.Map;
 
-    private final EndpointConfig config;
-    private final UpstreamService upstreamService;
+public final class UpstreamFactory {
 
-    DefaultEndpointService(EndpointConfig config) {
-        this.config = config;
-        upstreamService = UpstreamFactory.instance().get(config.upstream());
+    private static final UpstreamFactory INSTANCE = new UpstreamFactory();
+
+    public static UpstreamFactory instance() {
+        return INSTANCE;
     }
 
-    @Override
-    public UpstreamService upstream() {
-        return upstreamService;
+    private final Map<UpstreamConfig, UpstreamService> upstreams = new HashMap<>();
+
+    private UpstreamFactory() {}
+
+    public UpstreamService get(UpstreamConfig config) {
+        requireNonNull(config, "config");
+
+        final UpstreamService upstreamService = upstreams.get(config);
+        if (upstreamService != null) {
+            return upstreamService;
+        }
+
+        final UpstreamService newUpstreamService = UpstreamService.of(config);
+        upstreams.put(config, newUpstreamService);
+        return newUpstreamService;
     }
 }
